Refactor UI overlays through controller abstraction
