+++ PATCHLOG.txt
@@ -34,3 +34,4 @@ Spectra App — Patch Log (append-only)
 - v1.2.1h: Outline SpecViz-inspired ingestion, viewer, and plugin improvements so planning collateral guides upcoming parity work.
 - v1.2.1i: Normalise local uploads to Spectrum1D objects, validate flux/wavelength units with specutils, and extend ingestion regression coverage.
 - v1.2.1j: Capture image overlay statistics during local ingest and surface the richer metadata in the viewer summary.
+- v1.2.1k: Extract overlay/session helpers into a reusable controller, refactor the Streamlit overlay tab to use it, and document a CLI smoke demo with new controller tests.
