+++ app/version.json
@@ -1,5 +1,5 @@
 {
-  "version": "v1.2.1i",
-  "date_utc": "2025-10-23T00:00:00Z",
-  "summary": "Normalise local ingestion payloads to Spectrum1D objects and surface unit coercion errors."
+  "version": "v1.2.1k",
+  "date_utc": "2025-10-24T00:00:00Z",
+  "summary": "Factor overlay workspace logic into a reusable controller with CLI validation and new regression coverage."
 }
