+++ docs/ai_log/2025-10-24.md
@@ -1,0 +1,16 @@
+# AI Log — 2025-10-24
+
+## Tasking — Overlay controller extraction
+- Decouple overlay/session logic from Streamlit so alternative UIs can reuse the same workspace behaviour.
+
+## Actions & Decisions
+- Implemented `WorkspaceController` with overlay, viewport, differential, and similarity helpers while providing a thin Streamlit adapter. 【F:app/ui/controller.py†L1-L602】
+- Refactored `app/ui/main.py` to delegate overlay rendering, differential exports, and similarity setup to the controller. 【F:app/ui/main.py†L1621-L1900】
+- Added a CLI harness that exercises the controller without Streamlit and created controller-focused tests. 【F:app/controller_demo.py†L1-L74】【F:tests/ui/test_controller_module.py†L1-L66】
+
+## Verification
+- `pytest tests/ui/test_controller_module.py` 【90219d†L1-L6】
+- `python -m app.controller_demo` 【1cc0cf†L1-L10】
+
+## Docs Consulted
+- Internal controller design history (`docs/atlas/brains.md`). 【F:docs/atlas/brains.md†L1-L4】
