+++ docs/atlas/brains.md
@@ -1,3 +1,8 @@
+# Overlay controller extraction — 2025-10-24
+- Moved overlay/session helpers into `WorkspaceController` so Streamlit widgets call a reusable API and future UIs can reuse the same logic. 【F:app/ui/controller.py†L1-L602】
+- Updated the overlay tab to delegate to the controller for viewport management, figure rendering, differential exports, and similarity prep. 【F:app/ui/main.py†L1621-L1900】
+- Added a CLI harness that drives the controller without Streamlit to validate the abstraction ahead of alternate front ends. 【F:app/controller_demo.py†L1-L74】
+
 # Spectrum1D ingestion normalisation — 2025-10-5
 - Convert local ingest payloads into `Spectrum1D` objects with Astropy quantity checks so unit coercion errors surface before overlays are added, while skipping conversion for image products. 【F:app/utils/local_ingest.py†L300-L386】【F:app/utils/local_ingest.py†L597-L772】
 - Added specutils as a dependency and broadened ingestion tests to assert Spectrum1D availability and unit validation for base and additional traces. 【F:requirements.txt†L1-L11】【F:tests/server/test_local_ingest.py†L1-L186】【F:tests/server/test_local_ingest.py†L230-L321】
