+++ docs/patch_notes/v1.2.1k.md
@@ -1,0 +1,10 @@
+# v1.2.1k — Controller extraction for overlay workspace
+
+## Highlights
+- Factored overlay, viewport, and differential helpers into a new `WorkspaceController` module so Streamlit code delegates to pure-Python operations. 【F:app/ui/controller.py†L1-L602】
+- Refactored the Streamlit overlay tab to use the controller API and added a CLI demo that renders spectra without Streamlit. 【F:app/ui/main.py†L1621-L1900】【F:app/controller_demo.py†L1-L74】
+- Added focused controller unit tests covering figure and similarity preparation logic. 【F:tests/ui/test_controller_module.py†L1-L66】
+
+## Verification
+- `pytest tests/ui/test_controller_module.py` 【90219d†L1-L6】
+- `python -m app.controller_demo` 【1cc0cf†L1-L10】
